refactor(taskflow): extract app.nav + finalize Action split
